requirements.txt
requests
SQLAlchemy
mysql-connector-python
python-dotenv
pandas
